requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
altair==4.2.2
asttokens==2.2.1
async-timeout==4.0.3
attr==0.3.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
blinker==1.6.2
boltons==23.0.0
Brotli==1.0.9
brotlicffi==1.0.9.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
configparser==5.3.0
contextlib2==21.6.0
cryptography==40.0.2
Cython==0.29.34
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
dl==0.1.0
dnspython==2.4.2
docutils==0.20.1
entrypoints==0.4
executing==1.2.0
frozenlist==1.4.0
funcsigs==1.0.2
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
HTMLParser==0.0.2
httplib2==0.22.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
ipykernel==6.25.1
ipython==8.13.2
ipywidgets==8.0.6
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
keyring==23.13.1
langchain==0.0.173
lockfile==0.12.2
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mock==5.0.2
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numexpr==2.8.5
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
ordereddict==1.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==3.1.1
pypdf==3.15.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
railroad==0.5.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.6
sphinxcontrib-devhelp==1.0.4
sphinxcontrib-htmlhelp==2.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.5
sphinxcontrib-serializinghtml==1.1.7
SQLAlchemy==2.0.19
stack-data==0.6.2
streamlit==1.24.0
streamlit-chat==0.0.2.2
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
truststore==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
widgetsnbextension==4.0.8
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.16.2
